requirements.txt
yfinance
pandas
streamlit
plotly
sqlalchemy
psycopg2-binary
requests
newsapi-python
python-dotenv
lxml